# Reworks FluxPublish internals to relay on predictable state machine