--- a/reactor-core/src/main/java/reactor/core/publisher/FluxPublish.java
+++ b/reactor-core/src/main/java/reactor/core/publisher/FluxPublish.java
@@ -19,7 +19,6 @@ package reactor.core.publisher;
 import java.util.Objects;
 import java.util.Queue;
 import java.util.concurrent.CancellationException;
-import java.util.concurrent.atomic.AtomicIntegerFieldUpdater;
 import java.util.concurrent.atomic.AtomicLongFieldUpdater;
 import java.util.concurrent.atomic.AtomicReferenceFieldUpdater;
 import java.util.function.Consumer;
@@ -62,6 +61,8 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 	 */
 	final boolean resetUponSourceTermination;
 
+	final @Nullable Consumer<? super T> onDiscardHook;
+
 	volatile PublishSubscriber<T> connection;
 	@SuppressWarnings("rawtypes")
 	static final AtomicReferenceFieldUpdater<FluxPublish, PublishSubscriber> CONNECTION =
@@ -72,7 +73,8 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 	FluxPublish(Flux<? extends T> source,
 			int prefetch,
 			Supplier<? extends Queue<T>> queueSupplier,
-			boolean resetUponSourceTermination) {
+			boolean resetUponSourceTermination,
+			@Nullable Consumer<? super T> onDiscardHook) {
 		if (prefetch <= 0) {
 			throw new IllegalArgumentException("bufferSize > 0 required but it was " + prefetch);
 		}
@@ -80,6 +82,7 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 		this.prefetch = prefetch;
 		this.queueSupplier = Objects.requireNonNull(queueSupplier, "queueSupplier");
 		this.resetUponSourceTermination = resetUponSourceTermination;
+		this.onDiscardHook = onDiscardHook;
 	}
 
 	@Override
@@ -98,7 +101,7 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 				s = u;
 			}
 
-			doConnect = s.tryConnect();
+ 			doConnect = s.tryConnect();
 			break;
 		}
 
@@ -130,12 +133,15 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 			if (c.add(inner)) {
 				if (inner.isCancelled()) {
 					c.remove(inner);
-				} else {
+				}
+				else {
 					inner.parent = c;
 				}
-				c.drain();
+
+				c.drainFromInner();
 				break;
-			} else if (!this.resetUponSourceTermination) {
+			}
+			else if (!this.resetUponSourceTermination) {
 				if (c.error != null) {
 					inner.actual.onError(c.error);
 				} else {
@@ -168,12 +174,7 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 
 		final FluxPublish<T> parent;
 
-		volatile Subscription s;
-		@SuppressWarnings("rawtypes")
-		static final AtomicReferenceFieldUpdater<PublishSubscriber, Subscription> S =
-				AtomicReferenceFieldUpdater.newUpdater(PublishSubscriber.class,
-						Subscription.class,
-						"s");
+		Subscription s;
 
 		volatile PubSubInner<T>[] subscribers;
 
@@ -183,16 +184,10 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 						PubSubInner[].class,
 						"subscribers");
 
-		volatile int wip;
+		volatile long state;
 		@SuppressWarnings("rawtypes")
-		static final AtomicIntegerFieldUpdater<PublishSubscriber> WIP =
-				AtomicIntegerFieldUpdater.newUpdater(PublishSubscriber.class, "wip");
-
-		volatile int connected;
-		@SuppressWarnings("rawtypes")
-		static final AtomicIntegerFieldUpdater<PublishSubscriber> CONNECTED =
-				AtomicIntegerFieldUpdater.newUpdater(PublishSubscriber.class,
-						"connected");
+		static final AtomicLongFieldUpdater<PublishSubscriber> STATE =
+				AtomicLongFieldUpdater.newUpdater(PublishSubscriber.class, "state");
 
 		//notes: FluxPublish needs to distinguish INIT from CANCELLED in order to correctly
 		//drop values in case of an early connect() without any subscribers.
@@ -203,11 +198,11 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 		@SuppressWarnings("rawtypes")
 		static final PubSubInner[] TERMINATED = new PublishInner[0];
 
-		volatile Queue<T> queue;
+		Queue<T> queue;
 
 		int sourceMode;
 
-		volatile boolean   done;
+		boolean   done;
 
 		volatile Throwable error;
 
@@ -230,7 +225,9 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 
 		@Override
 		public void onSubscribe(Subscription s) {
-			if (Operators.setOnce(S, this, s)) {
+			if (Operators.validate(this.s, s)) {
+				this.s = s;
+
 				if (s instanceof Fuseable.QueueSubscription) {
 					@SuppressWarnings("unchecked") Fuseable.QueueSubscription<T> f =
 							(Fuseable.QueueSubscription<T>) s;
@@ -239,20 +236,42 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 					if (m == Fuseable.SYNC) {
 						sourceMode = m;
 						queue = f;
-						drain();
+						long previousState = markSubscriptionSetAndAddWork(this);
+						if (isCancelled(previousState)) {
+							s.cancel();
+							return;
+						}
+
+						if (hasWorkInProgress(previousState)) {
+							return;
+						}
+
+						drain(previousState, previousState | SUBSCRIPTION_SET_FLAG | 1);
 						return;
 					}
+
 					if (m == Fuseable.ASYNC) {
 						sourceMode = m;
 						queue = f;
-						s.request(Operators.unboundedOrPrefetch(prefetch));
+						long previousState = markSubscriptionSet(this);
+						if (isCancelled(previousState)) {
+							s.cancel();
+						}
+						else {
+							s.request(Operators.unboundedOrPrefetch(prefetch));
+						}
 						return;
 					}
 				}
 
 				queue = parent.queueSupplier.get();
-
-				s.request(Operators.unboundedOrPrefetch(prefetch));
+				long previousState = markSubscriptionSet(this);
+				if (isCancelled(previousState)) {
+					s.cancel();
+				}
+				else {
+					s.request(Operators.unboundedOrPrefetch(prefetch));
+				}
 			}
 		}
 
@@ -264,21 +283,35 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 				}
 				return;
 			}
-			if (sourceMode == Fuseable.ASYNC) {
-				drain();
-				return;
-			}
-
-			if (!queue.offer(t)) {
+			boolean isAsyncMode = sourceMode == Fuseable.ASYNC;
+			if (!isAsyncMode && !queue.offer(t)) {
 				Throwable ex = Operators.onOperatorError(s,
-						Exceptions.failWithOverflow(Exceptions.BACKPRESSURE_ERROR_QUEUE_FULL), t, currentContext());
+						Exceptions.failWithOverflow(Exceptions.BACKPRESSURE_ERROR_QUEUE_FULL),
+						t,
+						currentContext());
 				if (!Exceptions.addThrowable(ERROR, this, ex)) {
 					Operators.onErrorDroppedMulticast(ex, subscribers);
 					return;
 				}
 				done = true;
 			}
-			drain();
+
+			long previousState = addWork(this);
+
+			if (isFinalized(previousState)) {
+				clear();
+				return;
+			}
+
+			if (isTerminated(previousState) || isCancelled(previousState)) {
+				return;
+			}
+
+			if (hasWorkInProgress(previousState)) {
+				return;
+			}
+
+			drain(previousState, previousState + 1);
 		}
 
 		@Override
@@ -287,13 +320,22 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 				Operators.onErrorDroppedMulticast(t, subscribers);
 				return;
 			}
-			if (Exceptions.addThrowable(ERROR, this, t)) {
-				done = true;
-				drain();
-			}
-			else {
+
+			done = true;
+			if (!Exceptions.addThrowable(ERROR, this, t)) {
 				Operators.onErrorDroppedMulticast(t, subscribers);
 			}
+
+			long previousState = markTerminated(this);
+			if (isTerminated(previousState) || isCancelled(previousState)) {
+				return;
+			}
+
+			if (hasWorkInProgress(previousState)) {
+				return;
+			}
+
+			drain(previousState, (previousState | TERMINATED_FLAG) + 1);
 		}
 
 		@Override
@@ -302,7 +344,17 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 				return;
 			}
 			done = true;
-			drain();
+
+			long previousState = markTerminated(this);
+			if (isTerminated(previousState) || isCancelled(previousState)) {
+				return;
+			}
+
+			if (hasWorkInProgress(previousState)) {
+				return;
+			}
+
+			drain(previousState, (previousState | TERMINATED_FLAG) + 1);
 		}
 
 		@Override
@@ -311,19 +363,40 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 				return;
 			}
 			if (CONNECTION.compareAndSet(parent, this, null)) {
-				Operators.terminate(S, this);
-				if (WIP.getAndIncrement(this) != 0) {
+				long previousState = markCancelled(this);
+				if (isTerminated(previousState) || isCancelled(previousState)) {
 					return;
 				}
-				disconnectAction();
+
+				if (hasWorkInProgress(previousState)) {
+					return;
+				}
+
+				disconnectAction(previousState);
 			}
 		}
 
-		void disconnectAction() {
+		void clear() {
+			if (sourceMode == Fuseable.NONE) {
+				T t;
+				while ((t = queue.poll()) != null) {
+					Operators.onDiscard(t, currentContext());
+				}
+			}
+			else {
+				queue.clear();
+			}
+		}
+
+		void disconnectAction(long previousState) {
+			if (isSubscriptionSet(previousState)) {
+				this.s.cancel();
+				clear();
+			}
+
 			@SuppressWarnings("unchecked")
 			PubSubInner<T>[] inners = SUBSCRIBERS.getAndSet(this, CANCELLED);
 			if (inners.length > 0) {
-				queue.clear();
 				CancellationException ex = new CancellationException("Disconnected");
 
 				for (PubSubInner<T> inner : inners) {
@@ -392,16 +465,26 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 		}
 
 		boolean tryConnect() {
-			return connected == 0 && CONNECTED.compareAndSet(this, 0, 1);
+			long previousState = markConnected(this);
+
+			return !isConnected(previousState);
 		}
 
-		final void drain() {
-			if (WIP.getAndIncrement(this) != 0) {
+		final void drainFromInner() {
+			long previousState = addWorkIfSubscribed(this);
+
+			if (!isSubscriptionSet(previousState)) {
 				return;
 			}
 
-			int missed = 1;
+			if (hasWorkInProgress(previousState)) {
+				return;
+			}
 
+			drain(previousState, previousState + 1);
+		}
+
+		final void drain(long previousState, long expectedState) {
 			for (; ; ) {
 
 				boolean d = done;
@@ -514,16 +597,22 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 					}
 				}
 
-				missed = WIP.addAndGet(this, -missed);
-				if (missed == 0) {
-					break;
+				expectedState = markWorkDone(this, expectedState);
+				if (isCancelled(expectedState)) {
+					clearAndFinalize(this);
+					return;
+				}
+
+				if (!hasWorkInProgress(expectedState)) {
+					return;
 				}
 			}
 		}
 
 		boolean checkTerminated(boolean d, boolean empty) {
-			if (s == Operators.cancelledSubscription()) {
-				disconnectAction();
+			long state = this.state;
+			if (isCancelled(state)) {
+				disconnectAction(state);
 				return true;
 			}
 			if (d) {
@@ -578,9 +667,193 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 
 		@Override
 		public boolean isDisposed() {
-			return s == Operators.cancelledSubscription() || done;
+			long state = this.state;
+			return isTerminated(state) || isCancelled(state);
 		}
 
+		static void clearAndFinalize(PublishSubscriber<?> instance) {
+			for (; ; ) {
+				final long state = instance.state;
+
+				if (isFinalized(state)) {
+					instance.clear();
+					return;
+				}
+
+				if (isSubscriptionSet(state)) {
+					instance.clear();
+				}
+
+				if (STATE.compareAndSet(
+						instance, state,
+						(state & ~WORK_IN_PROGRESS_MASK) | FINALIZED_FLAG)) {
+					break;
+				}
+			}
+		}
+
+		static long addWork(PublishSubscriber<?> instance) {
+			for (;;) {
+				long state = instance.state;
+
+				if (STATE.compareAndSet(instance, state, addWork(state))) {
+					return state;
+				}
+			}
+		}
+
+		static long addWorkIfSubscribed(PublishSubscriber<?> instance) {
+			for (;;) {
+				long state = instance.state;
+
+				if (!isSubscriptionSet(state)) {
+					return state;
+				}
+
+				if (STATE.compareAndSet(instance, state, addWork(state))) {
+					return state;
+				}
+			}
+		}
+
+		static long addWork(long state) {
+			if ((state & WORK_IN_PROGRESS_MASK) == WORK_IN_PROGRESS_MASK) {
+				return (state &~ WORK_IN_PROGRESS_MASK) | 1;
+			}
+			else {
+				return state + 1;
+			}
+		}
+
+		static long markTerminated(PublishSubscriber<?> instance) {
+			for (;;) {
+				long state = instance.state;
+
+				if (isCancelled(state) || isTerminated(state)) {
+					return state;
+				}
+
+				long nextState = addWork(state);
+				if (STATE.compareAndSet(instance, state, nextState | TERMINATED_FLAG)) {
+					return state;
+				}
+			}
+		}
+
+		static long markConnected(PublishSubscriber<?> instance) {
+			for (;;) {
+				long state = instance.state;
+
+				if (isConnected(state)) {
+					return state;
+				}
+
+				if (STATE.compareAndSet(instance, state, state | CONNECTED_FLAG)) {
+					return state;
+				}
+			}
+		}
+
+		static long markSubscriptionSet(PublishSubscriber<?> instance) {
+			for (;;) {
+				long state = instance.state;
+
+				if (isCancelled(state)) {
+					return state;
+				}
+
+				if (STATE.compareAndSet(instance, state, state | SUBSCRIPTION_SET_FLAG)) {
+					return state;
+				}
+			}
+		}
+
+		static long markSubscriptionSetAndAddWork(PublishSubscriber<?> instance) {
+			for (;;) {
+				long state = instance.state;
+
+				if (isCancelled(state)) {
+					return state;
+				}
+
+				long nextState = addWork(state);
+				if (STATE.compareAndSet(instance, state, nextState | SUBSCRIPTION_SET_FLAG)) {
+					return state;
+				}
+			}
+		}
+
+		static long markCancelled(PublishSubscriber<?> instance) {
+			for (;;) {
+				long state = instance.state;
+
+				if (isCancelled(state)) {
+					return state;
+				}
+
+				long nextState = addWork(state);
+				if (STATE.compareAndSet(instance, state, nextState | CANCELLED_FLAG)) {
+					return state;
+				}
+			}
+		}
+
+		static long markWorkDone(PublishSubscriber<?> instance, long expectedState) {
+			for (;;) {
+				long state = instance.state;
+
+				if (expectedState != state) {
+					return state;
+				}
+
+				long nextState = state & ~WORK_IN_PROGRESS_MASK;
+				if (STATE.compareAndSet(instance, state, nextState)) {
+					return nextState;
+				}
+			}
+		}
+
+		static boolean isConnected(long state) {
+			return (state & CONNECTED_FLAG) == CONNECTED_FLAG;
+		}
+
+		static boolean isFinalized(long state) {
+			return (state & FINALIZED_FLAG) == FINALIZED_FLAG;
+		}
+
+		static boolean isCancelled(long state) {
+			return (state & CANCELLED_FLAG) == CANCELLED_FLAG;
+		}
+
+		static boolean isTerminated(long state) {
+			return (state & TERMINATED_FLAG) == TERMINATED_FLAG;
+		}
+
+		static boolean isSubscriptionSet(long state) {
+			return (state & SUBSCRIPTION_SET_FLAG) == SUBSCRIPTION_SET_FLAG;
+		}
+
+		static boolean hasWorkInProgress(long state) {
+			return (state & WORK_IN_PROGRESS_MASK) > 0;
+		}
+
+		static final long FINALIZED_FLAG =
+				0b0010_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000L;
+
+		static final long CANCELLED_FLAG =
+				0b0010_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000L;
+
+		static final long TERMINATED_FLAG =
+				0b0100_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000L;
+
+		static final long SUBSCRIPTION_SET_FLAG =
+				0b0000_1000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000L;
+
+		static final long CONNECTED_FLAG =
+				0b0000_0100_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000L;
+
+		static final long WORK_IN_PROGRESS_MASK =
+				0b0000_0000_0000_0000_0000_0000_0000_0000_1111_1111_1111_1111_1111_1111_1111_1111L;
 	}
 
 	static abstract class PubSubInner<T> implements InnerProducer<T> {
@@ -649,7 +922,7 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 		void drainParent() {
 			PublishSubscriber<T> p = parent;
 			if (p != null) {
-				p.drain();
+				p.drainFromInner();
 			}
 		}
 
@@ -658,7 +931,7 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 			PublishSubscriber<T> p = parent;
 			if (p != null) {
 				p.remove(this);
-				p.drain();
+				p.drainFromInner();
 			}
 		}