--- a/reactor-core/src/main/java/reactor/core/publisher/FluxPublish.java
+++ b/reactor-core/src/main/java/reactor/core/publisher/FluxPublish.java
@@ -61,8 +61,6 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 	 */
 	final boolean resetUponSourceTermination;
 
-	final @Nullable Consumer<? super T> onDiscardHook;
-
 	volatile PublishSubscriber<T> connection;
 	@SuppressWarnings("rawtypes")
 	static final AtomicReferenceFieldUpdater<FluxPublish, PublishSubscriber> CONNECTION =
@@ -73,8 +71,7 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 	FluxPublish(Flux<? extends T> source,
 			int prefetch,
 			Supplier<? extends Queue<T>> queueSupplier,
-			boolean resetUponSourceTermination,
-			@Nullable Consumer<? super T> onDiscardHook) {
+			boolean resetUponSourceTermination) {
 		if (prefetch <= 0) {
 			throw new IllegalArgumentException("bufferSize > 0 required but it was " + prefetch);
 		}
@@ -82,7 +79,6 @@ final class FluxPublish<T> extends ConnectableFlux<T> implements Scannable {
 		this.prefetch = prefetch;
 		this.queueSupplier = Objects.requireNonNull(queueSupplier, "queueSupplier");
 		this.resetUponSourceTermination = resetUponSourceTermination;
-		this.onDiscardHook = onDiscardHook;
 	}
 
 	@Override