This PR introduces a consistent statemachine `FluxPublish`. With that approach we have clear state migration which gives us an option to decide what action to do given the previous observed volatile state. This approach is useful when we need to discard values and ensure every given value is discarded once.

Alternativelly we can stick to the `FluxPublishOn` approach although it does not guarantee that every value is discarded once so in edgecases we may discard twice.

This approach does not bring any significant performance degradation nor performance benefits: 

**Baseline:**

```

Benchmark                               (rangeSize)  Mode  Cnt     Score    Error  Units
FluxPublishBenchmark.measureThroughput            0  avgt    5   673.753 ±  5.887  ns/op
FluxPublishBenchmark.measureThroughput           10  avgt    5   680.763 ±  3.550  ns/op
FluxPublishBenchmark.measureThroughput         1000  avgt    5   794.625 ±  6.296  ns/op
FluxPublishBenchmark.measureThroughput       100000  avgt    5  1172.740 ± 10.412  ns/op
```

**With state machine:**

```
Benchmark                               (rangeSize)  Mode  Cnt     Score    Error  Units
FluxPublishBenchmark.measureThroughput            0  avgt    5   687.399 ±  6.392  ns/op
FluxPublishBenchmark.measureThroughput           10  avgt    5   687.485 ±  4.141  ns/op
FluxPublishBenchmark.measureThroughput         1000  avgt    5   789.263 ±  4.399  ns/op
FluxPublishBenchmark.measureThroughput       100000  avgt    5  1105.864 ± 12.900  ns/op
```