@chemicL 2023-08-07
As discussed, this PR addresses future discard functionality, while for the moment is a refactor that unifies multiple volatile fields into one state machine which gives guarantees we can take advantage of later.

@OlegDokuka please check spotless failure and I'll approve once the build passes.